Refactor to standard project structure
