+++ README.md
@@ -3,33 +3,28 @@
 ## Overview
 This is the RAS Party micro-service. See [Confluence] for further information.
 
+The API is specified [here](./API.md).
+
 
 ## Tests
 To run the tests with [tox], install tox (see link) then simply run the command `tox` in the root of the project.
 tox will create a unique virtualenv, run the unit tests with py.test, then run flake8 coverage.
 
-To install and run. Firstly create a virtualenv, then:
+To install and run: Create a (Python 3.5+) virtualenv, then:
 ``` bash
-pip3 install -r requirements.txt
-python3 run.py
+pip install -r requirements.txt
+python run.py
 ```
 
-and open your browser to here:
+To test the service is up:
 
 ```
-http://localhost:8080/party-api/1.0.4/ui/
-```
-
-Your Swagger definition lives here:
-
-```
-http://localhost:8080/party-api/1.0.4/swagger.json
+curl http://localhost:8080/party-api/v1/info
 ```
 
 ## Database
 
-The database will automatically be created when starting the application. It is also possible to drop/create the
-database via create_db.py. Run `python database.py` in the project root to see further guidance.
+The database will automatically be created when starting the application.
 
 
 [Confluence]: https://digitaleq.atlassian.net/wiki/display/RASB/Party
