+++ app.py
@@ -6,7 +6,7 @@ from ras_common_utils.ras_config.flask_extended import Flask
 from ras_common_utils.ras_database.ras_database import RasDatabase
 from ras_common_utils.ras_logger.ras_logger import configure_logger
 
-from swagger_server.controllers.gw_registration import make_registration_func, call_in_background
+from ras_party.controllers.gw_registration import make_registration_func, call_in_background
 
 """
 This is a duplicate of run.py, with minor modifications to support gunicorn execution.
@@ -36,7 +36,7 @@ def create_app(config):
 
 def initialise_db(app):
     # Initialise the database with the specified SQLAlchemy model
-    PartyDatabase = RasDatabase.make(model_paths=['swagger_server.models.models'])
+    PartyDatabase = RasDatabase.make(model_paths=['ras_party.models.models'])
     db = PartyDatabase('ras-party-db', app.config)
     # TODO: this isn't entirely safe, use a get_db() lazy initializer instead...
     app.db = db
