+++ ras_party/controllers/util.py
@@ -53,3 +53,8 @@ def flatten_keys(d, prefix=None):
             result.extend(flatten_keys(v, prefix='.'.join([prefix, k] if prefix else [k])))
 
     return result
+
+
+def build_url(template, config, *args):
+    url = template.format(config['scheme'], config['host'], config['port'], *args)
+    return url
