+++ ras_party/controllers/validate.py
@@ -1,7 +1,7 @@
 import itertools
 import uuid
 
-from swagger_server.controllers.util import flatten_keys
+from ras_party.controllers.util import flatten_keys
 
 
 class ValidatorBase:
