--- swagger_server/controllers/url_builder.py
@@ -1,3 +1,0 @@
-def build_url(template, config, *args):
-    url = template.format(config['scheme'], config['host'], config['port'], *args)
-    return url
