--- swagger_server/models/_models.py
@@ -1,3 +1,0 @@
-from swagger_server.models.models import Business, BusinessRespondentStatus, BusinessRespondent, \
-    RespondentStatus, Respondent, EnrolmentStatus, Enrolment, EnrolmentCodeStatus, EnrolmentCode, \
-    EnrolmentInvitationStatus, EnrolmentInvitation
