+++ test/fixtures/get_ce_by_id.py
