+++ test/test_util.py
@@ -1,6 +1,6 @@
 from unittest import TestCase
 
-from swagger_server.controllers.util import flatten_keys
+from ras_party.controllers.util import flatten_keys
 
 
 class TestUtil(TestCase):
