+++ tox.ini
@@ -7,7 +7,7 @@ deps=-r{toxinidir}/requirements.txt
 
 commands=
 ;   flake8 . --count    ; FIXME: the code generator produces code which fails linting checks
-   py.test --cov=swagger_server/controllers --cov-report html --cov-report term-missing \
+   py.test --cov=ras_party --cov-report html --cov-report term-missing \
       []
 
 [flake8]
